chore(comments): Convert comments JS to AMD
