+++ engine/lib/elgglib.php
@@ -1908,6 +1908,7 @@ function _elgg_init() {
 	
 	elgg_register_external_view('js/elgg/UserPicker.js', true);
 	elgg_register_simplecache_view('js/elgg/booted.js');
+	elgg_require_js('elgg/Comment');
 	elgg_require_js('elgg/booted');
 
 	elgg_register_js('elgg.friendspicker', 'js/lib/ui.friends_picker.js');
